refactor(cos): 文件调用改为路径+参数的统一形式
